refactor: dummy item companion object로 이동
